+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+NEXT:
+ - Uses `libwebp` over a `libwebp-jni` instead of OpenCV to create webp images
+
 1.4.0:
  - Use new format for PupilIDs with picture and scaled picture externalized from the VC
  - revamped revocation list caching